requirements/production.txt
numpy==2.1.1
python-dotenv==1.0.1
redis==5.1.0
typer==0.12.5